+++ node/lib/retire.js
@@ -4,7 +4,7 @@
  */
 
 var exports = exports || {};
-exports.version = '5.6.0';
+exports.version = '5.7.0';
 
 function isDefined(o) {
   return typeof o !== 'undefined';