+++ node/src/reporters/cyclonedx-1_7-json.ts
@@ -1,3 +1,3 @@
 import { cycloneDXJSONLogger } from './cyclonedx-json-builder';
 
-export default cycloneDXJSONLogger('1.6');
+export default cycloneDXJSONLogger('1.7');